refactor: Use mappers in ProductService and UserService refactor: Update ProductController and UserController to use ResponseEntity and services feat: Add custom exceptions for Product and User

+++ src/main/java/com/pedro_a10/LojaAPI/service/ProductService.java
@@ -1,13 +1,17 @@
 package com.pedro_a10.LojaAPI.service;
 
+import com.pedro_a10.LojaAPI.dto.productdto.ProductRequestDTO;
+import com.pedro_a10.LojaAPI.dto.productdto.ProductResponseDTO;
 import com.pedro_a10.LojaAPI.entity.Product;
 import com.pedro_a10.LojaAPI.enums.ProductType;
+import com.pedro_a10.LojaAPI.exceptions.ProductInvalidQuantityException;
+import com.pedro_a10.LojaAPI.exceptions.ProductNotFoundException;
+import com.pedro_a10.LojaAPI.mapper.ProductMapper;
 import com.pedro_a10.LojaAPI.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class ProductService {
@@ -15,27 +19,43 @@ public class ProductService {
   @Autowired
   ProductRepository productRepository;
 
-  public Optional<Product> findById(Long id) {
-    return productRepository.findById(id);
+  @Autowired
+  ProductMapper productMapper;
+
+  public ProductResponseDTO findById(Long id) {
+    Product product = productRepository.findById(id)
+      .orElseThrow(() -> new ProductNotFoundException("Product not found with id: " + id));
+    return productMapper.toResponseDTO(product);
   }
 
-  public Optional<Product> findByName(String name) {
-    return productRepository.findByName(name);
+  public ProductResponseDTO findByName(String name) {
+    Product productName = productRepository.findByName(name)
+      .orElseThrow(() -> new ProductNotFoundException("Product not found with name: " + name));
+    return productMapper.toResponseDTO(productName);
   }
 
-  public List<Product> findByPriceInCents(Integer priceInCents) {
-    return productRepository.findByPriceInCents(priceInCents);
+  public List<ProductResponseDTO> findByPriceInCents(Integer priceInCents) {
+    return productMapper.toResponseDTOList(productRepository.findByPriceInCents(priceInCents));
   }
 
-  public List<Product> findAllByType(ProductType type) {
-    return productRepository.findAllByType(type);
+  public List<ProductResponseDTO> findAllByType(ProductType type) {
+    return productMapper.toResponseDTOList(productRepository.findAllByType(type));
   }
 
-  public Product createProduct(Product product) {
-    return productRepository.save(product);
+  public ProductResponseDTO createProduct(ProductRequestDTO productRequestDTO) {
+    if (productRequestDTO.getQuantity() <= 0){
+      throw new ProductInvalidQuantityException("Is not possible create products with 0 or below quantity" + productRequestDTO.getQuantity());
+    }
+
+    Product products = productMapper.toEntity(productRequestDTO);
+    Product saveProducts = productRepository.save(products);
+    return productMapper.toResponseDTO(saveProducts);
   }
 
   public void deleteById(Long id) {
+    if (!productRepository.existsById(id)) {
+      throw new ProductNotFoundException("Product not found with id " + id);
+    }
     productRepository.deleteById(id);
   }
 }
